refactored e2e test and helpers
