+++ src/handlers/int-test-support/helpers/testDataHelper.ts
@@ -128,34 +128,3 @@ export interface TestData {
   billingPriceFormatted: string;
   transactionPriceFormatted: string;
 }
-
-export const getEmailAddresses = async (): Promise<{
-  sourceEmail: string;
-  toEmail: string;
-}> => {
-  const prefix = resourcePrefix();
-  const extractedEnvValue = prefix.split("-").pop();
-
-  if (extractedEnvValue === undefined) {
-    throw new Error("Env is undefined");
-  }
-  let sourceEmail = "";
-  let toEmail = "";
-  if (
-    extractedEnvValue.includes("dev") ||
-    extractedEnvValue.includes("build")
-  ) {
-    sourceEmail = `no-reply@btm.${extractedEnvValue}.account.gov.uk`;
-    toEmail = `vendor1_invoices@btm.${extractedEnvValue}.account.gov.uk`;
-  } else if (
-    extractedEnvValue?.includes("staging") ||
-    extractedEnvValue?.includes("integration")
-  ) {
-    sourceEmail = `no-reply@btm.${extractedEnvValue}.account.gov.uk`;
-    toEmail = (await getE2ETestConfig()).parser_0_toEmailId;
-    console.log("EmailId from config:", toEmail);
-  } else {
-    console.error(`Email domains are not exists for the given ${prefix}`);
-  }
-  return { sourceEmail, toEmail };
-};
